Create test core gates controlled by file
