+++ src/qibo/tests/test_gates.py
@@ -15,181 +15,6 @@ _DEVICE_BACKENDS = [("custom", None), ("matmuleinsum", None),
 
 
 @pytest.mark.parametrize("backend", _BACKENDS)
-def test_multicontrol_xgate(backend):
-    """Check that fallback method for X works for one or two controls."""
-    original_backend = qibo.get_backend()
-    qibo.set_backend(backend)
-    c1 = Circuit(3)
-    c1.add(gates.X(0))
-    c1.add(gates.X(2).controlled_by(0))
-    final_state = c1.execute()
-    c2 = Circuit(3)
-    c2.add(gates.X(0))
-    c2.add(gates.CNOT(0, 2))
-    target_state = c2.execute()
-    np.testing.assert_allclose(final_state, target_state)
-
-    c1 = Circuit(3)
-    c1.add(gates.X(0))
-    c1.add(gates.X(2))
-    c1.add(gates.X(1).controlled_by(0, 2))
-    final_state = c1.execute()
-    c2 = Circuit(3)
-    c2.add(gates.X(0))
-    c2.add(gates.X(2))
-    c2.add(gates.TOFFOLI(0, 2, 1))
-    target_state = c2.execute()
-    np.testing.assert_allclose(final_state, target_state)
-    qibo.set_backend(original_backend)
-
-
-@pytest.mark.parametrize(("backend", "accelerators"), _DEVICE_BACKENDS)
-def test_multicontrol_xgate_more_controls(backend, accelerators):
-    """Check that fallback method for X works for more than two controls."""
-    original_backend = qibo.get_backend()
-    qibo.set_backend(backend)
-    c = Circuit(4, accelerators)
-    c.add(gates.X(0))
-    c.add(gates.X(1))
-    c.add(gates.X(2))
-    c.add(gates.X(3).controlled_by(0, 1, 2))
-    c.add(gates.X(0))
-    c.add(gates.X(2))
-    final_state = c.execute()
-
-    c = Circuit(4)
-    c.add(gates.X(1))
-    c.add(gates.X(3))
-    target_state = c.execute()
-
-    np.testing.assert_allclose(final_state, target_state)
-    qibo.set_backend(original_backend)
-
-
-@pytest.mark.parametrize("backend", _BACKENDS)
-def test_controlled_u1(backend):
-    """Check controlled U1 and fallback to CU1."""
-    original_backend = qibo.get_backend()
-    qibo.set_backend(backend)
-    theta = 0.1234
-
-    c = Circuit(3)
-    c.add(gates.X(0))
-    c.add(gates.X(1))
-    c.add(gates.X(2))
-    c.add(gates.U1(2, theta).controlled_by(0, 1))
-    c.add(gates.X(0))
-    c.add(gates.X(1))
-    final_state = c.execute()
-    target_state = np.zeros_like(final_state)
-    target_state[1] = np.exp(1j * theta)
-    np.testing.assert_allclose(final_state, target_state)
-
-    gate = gates.U1(0, theta).controlled_by(1)
-    assert gate.__class__.__name__ == "CU1"
-    qibo.set_backend(original_backend)
-
-
-@pytest.mark.parametrize("backend", _BACKENDS)
-def test_controlled_u2(backend):
-    """Check controlled by U2."""
-    original_backend = qibo.get_backend()
-    qibo.set_backend(backend)
-    phi = 0.1234
-    lam = 0.4321
-
-    c = Circuit(3)
-    c.add([gates.X(0), gates.X(1)])
-    c.add(gates.U2(2, phi, lam).controlled_by(0, 1))
-    c.add([gates.X(0), gates.X(1)])
-    final_state = c()
-
-    c = Circuit(3)
-    c.add([gates.X(0), gates.X(1)])
-    c.add(gates.U2(2, phi, lam))
-    c.add([gates.X(0), gates.X(1)])
-    target_state = c()
-    np.testing.assert_allclose(final_state, target_state)
-
-    # for coverage
-    gate = gates.CU2(0, 1, phi, lam)
-    assert gate.parameters == (phi, lam)
-    qibo.set_backend(original_backend)
-
-
-@pytest.mark.parametrize("backend", _BACKENDS)
-def test_controlled_u3(backend):
-    """Check controlled U3 fall backs to CU3."""
-    original_backend = qibo.get_backend()
-    qibo.set_backend(backend)
-    theta = 0.1
-    phi = 0.1234
-    lam = 0.4321
-
-    initial_state = utils.random_numpy_state(2)
-    c = Circuit(2)
-    c.add(gates.U3(1, theta, phi, lam).controlled_by(0))
-    final_state = c(np.copy(initial_state))
-    assert c.queue[0].__class__.__name__ == "CU3"
-
-    c = Circuit(2)
-    c.add(gates.CU3(0, 1, theta, phi, lam))
-    target_state = c(np.copy(initial_state))
-    np.testing.assert_allclose(final_state, target_state)
-
-    # for coverage
-    gate = gates.U3(0, theta, phi, lam)
-    assert gate.parameters == (theta, phi, lam)
-    qibo.set_backend(original_backend)
-
-
-@pytest.mark.parametrize(("backend", "accelerators"), _DEVICE_BACKENDS)
-def test_cu1(backend, accelerators):
-    """Check CU1 gate is working properly on |11>."""
-    original_backend = qibo.get_backend()
-    qibo.set_backend(backend)
-    theta = 0.1234
-
-    c = Circuit(2, accelerators)
-    c.add(gates.X(0))
-    c.add(gates.X(1))
-    c.add(gates.CU1(0, 1, theta))
-    final_state = c.execute()
-
-    phase = np.exp(1j * theta)
-    target_state = np.zeros_like(final_state)
-    target_state[3] = phase
-    np.testing.assert_allclose(final_state, target_state)
-    qibo.set_backend(original_backend)
-
-
-def test_controlled_by_fsim(backend, accelerators):
-    """Check ``controlled_by`` for fSim gate."""
-    original_backend = qibo.get_backend()
-    qibo.set_backend(backend)
-    theta = 0.1234
-    phi = 0.4321
-
-    c = Circuit(6, accelerators)
-    c.add((gates.H(i) for i in range(6)))
-    c.add(gates.fSim(5, 3, theta, phi).controlled_by(0, 2, 1))
-    final_state = c.execute()
-
-    target_state = np.ones_like(final_state) / np.sqrt(2 ** 6)
-    rotation = np.array([[np.cos(theta), -1j * np.sin(theta)],
-                         [-1j * np.sin(theta), np.cos(theta)]])
-    matrix = np.eye(4, dtype=target_state.dtype)
-    matrix[1:3, 1:3] = rotation
-    matrix[3, 3] = np.exp(-1j * phi)
-    ids = [56, 57, 60, 61]
-    target_state[ids] = matrix.dot(target_state[ids])
-    ids = [58, 59, 62, 63]
-    target_state[ids] = matrix.dot(target_state[ids])
-    np.testing.assert_allclose(final_state, target_state)
-    qibo.set_backend(original_backend)
-
-
-@pytest.mark.parametrize("backend", _BACKENDS)
 def test_generalized_fsim_error(backend):
     """Check GenerelizedfSim gate raises error for wrong unitary shape."""
     original_backend = qibo.get_backend()
@@ -202,114 +27,8 @@ def test_generalized_fsim_error(backend):
     qibo.set_backend(original_backend)
 
 
-@pytest.mark.parametrize("backend", _BACKENDS)
-def test_doubly_controlled_by_rx_no_effect(backend):
-    original_backend = qibo.get_backend()
-    qibo.set_backend(backend)
-    theta = 0.1234
-
-    c = Circuit(3)
-    c.add(gates.X(0))
-    c.add(gates.RX(2, theta).controlled_by(0, 1))
-    c.add(gates.X(0))
-    final_state = c.execute()
-
-    target_state = np.zeros_like(final_state)
-    target_state[0] = 1.0
-
-    np.testing.assert_allclose(final_state, target_state)
-    qibo.set_backend(original_backend)
-
-
-@pytest.mark.parametrize(("backend", "accelerators"), _DEVICE_BACKENDS)
-def test_doubly_controlled_by_rx(backend, accelerators):
-    original_backend = qibo.get_backend()
-    qibo.set_backend(backend)
-    theta = 0.1234
-
-    c = Circuit(3, accelerators)
-    c.add(gates.RX(2, theta))
-    target_state = c.execute()
-
-    c = Circuit(3)
-    c.add(gates.X(0))
-    c.add(gates.X(1))
-    c.add(gates.RX(2, theta).controlled_by(0, 1))
-    c.add(gates.X(0))
-    c.add(gates.X(1))
-    final_state = c.execute()
-
-    np.testing.assert_allclose(final_state, target_state)
-    qibo.set_backend(original_backend)
-
-
-@pytest.mark.parametrize("backend", _BACKENDS)
-def test_controlled_by_swap_small(backend):
-    """Check controlled SWAP using controlled by for ``nqubits=3``."""
-    original_backend = qibo.get_backend()
-    qibo.set_backend(backend)
-    c = Circuit(3)
-    c.add(gates.RX(1, theta=0.1234))
-    c.add(gates.RY(2, theta=0.4321))
-    c.add(gates.SWAP(1, 2).controlled_by(0))
-    final_state = c.execute()
-    c = Circuit(3)
-    c.add(gates.RX(1, theta=0.1234))
-    c.add(gates.RY(2, theta=0.4321))
-    target_state = c.execute()
-    np.testing.assert_allclose(final_state, target_state)
-
-    c = Circuit(3)
-    c.add(gates.X(0))
-    c.add(gates.RX(1, theta=0.1234))
-    c.add(gates.RY(2, theta=0.4321))
-    c.add(gates.SWAP(1, 2).controlled_by(0))
-    c.add(gates.X(0))
-    final_state = c.execute()
-    c = Circuit(3)
-    c.add(gates.RX(1, theta=0.1234))
-    c.add(gates.RY(2, theta=0.4321))
-    c.add(gates.SWAP(1, 2))
-    target_state = c.execute()
-    np.testing.assert_allclose(final_state, target_state)
-    qibo.set_backend(original_backend)
-
-
-@pytest.mark.parametrize("backend", _BACKENDS)
-def test_controlled_by_swap(backend):
-    """Check controlled SWAP using controlled by for ``nqubits=4``."""
-    original_backend = qibo.get_backend()
-    qibo.set_backend(backend)
-    c = Circuit(4)
-    c.add(gates.RX(2, theta=0.1234))
-    c.add(gates.RY(3, theta=0.4321))
-    c.add(gates.SWAP(2, 3).controlled_by(0))
-    final_state = c.execute()
-    c = Circuit(4)
-    c.add(gates.RX(2, theta=0.1234))
-    c.add(gates.RY(3, theta=0.4321))
-    target_state = c.execute()
-    np.testing.assert_allclose(final_state, target_state)
-
-    c = Circuit(4)
-    c.add(gates.X(0))
-    c.add(gates.RX(2, theta=0.1234))
-    c.add(gates.RY(3, theta=0.4321))
-    c.add(gates.SWAP(2, 3).controlled_by(0))
-    c.add(gates.X(0))
-    final_state = c.execute()
-    c = Circuit(4)
-    c.add(gates.RX(2, theta=0.1234))
-    c.add(gates.RY(3, theta=0.4321))
-    c.add(gates.SWAP(2, 3))
-    target_state = c.execute()
-    np.testing.assert_allclose(final_state, target_state)
-    qibo.set_backend(original_backend)
-
-
-@pytest.mark.parametrize(("backend", "accelerators"), _DEVICE_BACKENDS)
+@pytest.mark.parametrize("backend,accelerators", _DEVICE_BACKENDS)
 def test_doubly_controlled_by_swap(backend, accelerators):
-    """Check controlled SWAP using controlled by two qubits."""
     original_backend = qibo.get_backend()
     qibo.set_backend(backend)
     c = Circuit(4, accelerators)
